Refactor code into cmd/ and pkg/ directories.
